Add WebAssembly support and some browser examples.
The `zstd` crate we used for reading and writing zstd does not support wasm.

The (pure rust) `ruzstd` supports wasm, but only supports reading.

So we now use `ruzstd` for reading, and `zstd` for writing. So, writing
is not supported on wasm, only reading.

Introduces a javascript API in the new geomedea-wasm package (based on `wasm-pack new geomedea-wasm`)